Documents/Git/survey_mgnt/requirements_colab.txt
# Requirements for Google Colab Survey Deduplication with Sheets Export
# Install using: !pip install -r requirements_colab.txt

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Snowflake connectivity
snowflake-connector-python>=3.0.0
sqlalchemy>=1.4.0

# Google Sheets integration
gspread>=5.7.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0

# Additional utilities
openpyxl>=3.0.9  # For Excel export
requests>=2.28.0
urllib3>=1.26.0

# Optional: Enhanced text processing
# scikit-learn>=1.1.0  # Uncomment if using TF-IDF features
# sentence-transformers>=2.2.0  # Uncomment if using semantic similarity 